Started Adding SpinIntake Command
Added the SpinIntake command which will spin the intake while the trigger on the controller is held, as well as all of the code for it in the robot container. Please note that the {button}.whileHeld() has been depricated, and .whileTrue must be used instead. (look at deprication notes in VSCode)