Added Compressor Class
Added a baseline class for the compressor as a subsystem, such that it can be turned on and off at will. The main purpose in doing this is to get maximum efficiency when driving. i.e. we are driving and the compressor kicks on which soaks up current from the battery and either slows us down, or browns us out.