Added a bare bones autonomous command
I added a new autonomous command with just a bare bones skeleton structure to be filled in with meat. Please read the notes for 2 / 18 for additional information. I added getter and reset classes for our nav board as well, mainly to be used in our PID loop, but can also be used for driver control, so I put it in the drivetrain subsystem, just like our enocder values for encoders that don't exist on the robot.