Auto works. Read Desc.
Auto works. The issue where the drivetrain was jittering was an issue with the scheduler running additional tasks in the background, which were somehow overriding the values for the drivetrain. To fix this, we can use a time based for loop which will isolate the program until the the conditions are met.